refactor: move source to src/ as .mjs
Align with the JS-ESM archetype convention adopted across the Wave 1-2
leaves (src/ + explicit .mjs extension).

- lib/ -> src/; index.js -> index.mjs; index.d.ts -> index.d.mts
  (NodeNext resolves a .mjs import's types from .d.mts, not .d.ts)
- repoint exports/types, vitest coverage include, tests, benchmarks
