+++ package.json
@@ -14,16 +14,16 @@
         "url": "git://github.com/hapijs/bourne.git"
     },
     "files": [
-        "lib",
+        "src",
         "API.md",
         "README.md"
     ],
     "type": "module",
-    "types": "lib/index.d.ts",
+    "types": "src/index.d.mts",
     "exports": {
         ".": {
-            "types": "./lib/index.d.ts",
-            "default": "./lib/index.js"
+            "types": "./src/index.d.mts",
+            "default": "./src/index.mjs"
         }
     },
     "scripts": {
